Fixed usage order and option parsing.
The usage statement required options to follow arguments, eg.
   gbss_1_reg.sh <output> [options]
but getopts expects options to be parsed first, and stops when
it gets to the first thing that isn't an option. Thus, getopts
was stopping immediately and never parsing the options.

In this commit I've reworked the usage help and parsing.

I've also renamed the argument from 'input_directory' to what I
think is the proper name, 'output_directory'. :-)